refactor(dashboard): address PR review - split components into files, fix hooks
- Move LogEntryRow, JobViewerHeader, GroupResultsList, JobLogsSection, JobCard
  each to their own file (per review: separate file)
- Move useGroupsPageData to lib/hooks/storage/useGroupsPageData.ts
- Change useGroupsPageData params to object for safe destructuring (per review)
- Add Mutation suffix to all mutation variables in GroupsPage (per review)
- Add formatLogTime shared helper to lib/utils; use formatDateTime in JobCard
  replacing local formatTimestamp (per review: use shared helper)
- Fix react-hooks/exhaustive-deps: wrap query.data ?? [] in useMemo
- Add max-lines: 200 rule for tsx/jsx component files
- Fix test file max-lines-per-function override order in extension eslint config
- Suppress react-hooks/incompatible-library for VirtualPostList (TanStack false positive)
