+++ extension/entrypoints/dashboard/components/GroupsPage.tsx
@@ -1,55 +1,15 @@
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
 import {
   useBulkDeleteGroups,
   useBulkUpdateGroups,
   useDeleteGroup,
-  useGroups,
   useScanGroupsList,
   useUpdateGroup,
 } from '@/lib/hooks/storage/useGroups';
-import { useSubscriptions } from '@/lib/hooks/storage/useSubscriptions';
+import { useGroupsPageData } from '@/lib/hooks/storage/useGroupsPageData';
 import { BulkActionsBar } from './BulkActionsBar';
 import { GroupsPageHeader } from './GroupsPageHeader';
 import { GroupsTable } from './GroupsTable';
-
-function useGroupsPageData(
-  filterSubscriptionId: string | null,
-  searchQuery: string
-) {
-  const groupsQuery = useGroups();
-  const subscriptionsQuery = useSubscriptions();
-
-  const groups = groupsQuery.data ?? [];
-  const subscriptions = subscriptionsQuery.data ?? [];
-  const isLoading = groupsQuery.isLoading || subscriptionsQuery.isLoading;
-
-  const filteredGroups = useMemo(() => {
-    let result = groups;
-    if (filterSubscriptionId) {
-      result = result.filter((g) =>
-        filterSubscriptionId === 'unassigned'
-          ? g.subscriptionIds.length === 0
-          : g.subscriptionIds.includes(filterSubscriptionId)
-      );
-    }
-    if (searchQuery) {
-      const query = searchQuery.toLowerCase();
-      result = result.filter((g) => g.name.toLowerCase().includes(query));
-    }
-    return result;
-  }, [groups, filterSubscriptionId, searchQuery]);
-
-  const stats = useMemo(() => {
-    const total = groups.length;
-    const unassigned = groups.filter(
-      (g) => g.subscriptionIds.length === 0
-    ).length;
-    const enabled = groups.filter((g) => g.enabled).length;
-    return { total, unassigned, enabled };
-  }, [groups]);
-
-  return { groups, subscriptions, isLoading, filteredGroups, stats };
-}
 
 export function GroupsPage() {
   const [selectedGroupIds, setSelectedGroupIds] = useState<Set<string>>(
@@ -62,18 +22,15 @@ export function GroupsPage() {
   const [bulkSubscriptionId, setBulkSubscriptionId] = useState('');
 
   const { subscriptions, isLoading, filteredGroups, stats } = useGroupsPageData(
-    filterSubscriptionId,
-    searchQuery
+    { filterSubscriptionId, searchQuery }
   );
 
-  // Mutations
-  const scanGroupsList = useScanGroupsList();
-  const updateGroup = useUpdateGroup();
-  const deleteGroup = useDeleteGroup();
-  const bulkUpdateGroups = useBulkUpdateGroups();
-  const bulkDeleteGroups = useBulkDeleteGroups();
+  const scanGroupsListMutation = useScanGroupsList();
+  const updateGroupMutation = useUpdateGroup();
+  const deleteGroupMutation = useDeleteGroup();
+  const bulkUpdateGroupsMutation = useBulkUpdateGroups();
+  const bulkDeleteGroupsMutation = useBulkDeleteGroups();
 
-  // Handle checkbox toggle
   function handleToggleSelection(groupId: string) {
     setSelectedGroupIds((prev) => {
       const next = new Set(prev);
@@ -86,7 +43,6 @@ export function GroupsPage() {
     });
   }
 
-  // Handle select all
   function handleSelectAll() {
     if (selectedGroupIds.size === filteredGroups.length) {
       setSelectedGroupIds(new Set());
@@ -95,11 +51,9 @@ export function GroupsPage() {
     }
   }
 
-  // Handle bulk assign to subscription
   function handleBulkAssign() {
     if (!bulkSubscriptionId || selectedGroupIds.size === 0) return;
-
-    bulkUpdateGroups.mutate(
+    bulkUpdateGroupsMutation.mutate(
       {
         groupIds: Array.from(selectedGroupIds),
         updates: { subscriptionIds: [bulkSubscriptionId], enabled: true },
@@ -113,27 +67,16 @@ export function GroupsPage() {
     );
   }
 
-  // Handle bulk enable/disable
   function handleBulkToggleEnabled(enabled: boolean) {
     if (selectedGroupIds.size === 0) return;
-
-    bulkUpdateGroups.mutate(
-      {
-        groupIds: Array.from(selectedGroupIds),
-        updates: { enabled },
-      },
-      {
-        onSuccess: () => {
-          setSelectedGroupIds(new Set());
-        },
-      }
+    bulkUpdateGroupsMutation.mutate(
+      { groupIds: Array.from(selectedGroupIds), updates: { enabled } },
+      { onSuccess: () => setSelectedGroupIds(new Set()) }
     );
   }
 
-  // Handle bulk delete
   function handleBulkDelete() {
     if (selectedGroupIds.size === 0) return;
-
     if (
       !confirm(
         `Delete ${selectedGroupIds.size} group(s)? This will also delete all posts from these groups.`
@@ -141,32 +84,29 @@ export function GroupsPage() {
     ) {
       return;
     }
-
-    bulkDeleteGroups.mutate(Array.from(selectedGroupIds), {
-      onSuccess: () => {
-        setSelectedGroupIds(new Set());
-      },
+    bulkDeleteGroupsMutation.mutate(Array.from(selectedGroupIds), {
+      onSuccess: () => setSelectedGroupIds(new Set()),
     });
   }
 
-  // Handle individual group toggle
   function handleToggleGroup(groupId: string, enabled: boolean) {
-    updateGroup.mutate({ id: groupId, updates: { enabled } });
+    updateGroupMutation.mutate({ id: groupId, updates: { enabled } });
   }
 
-  // Handle individual group assignment
   function handleAssignGroup(groupId: string, subscriptionId: string) {
     const subscriptionIds = subscriptionId ? [subscriptionId] : [];
     const enabled = !!subscriptionId;
-    updateGroup.mutate({ id: groupId, updates: { subscriptionIds, enabled } });
+    updateGroupMutation.mutate({
+      id: groupId,
+      updates: { subscriptionIds, enabled },
+    });
   }
 
-  // Handle individual group delete
   function handleDeleteGroup(groupId: string) {
     if (!confirm('Delete this group? This will also delete all its posts.')) {
       return;
     }
-    deleteGroup.mutate(groupId);
+    deleteGroupMutation.mutate(groupId);
   }
 
   if (isLoading) {
@@ -184,11 +124,10 @@ export function GroupsPage() {
         searchQuery={searchQuery}
         filterSubscriptionId={filterSubscriptionId}
         subscriptions={subscriptions}
-        scanGroupsList={scanGroupsList}
+        scanGroupsList={scanGroupsListMutation}
         onSearchChange={setSearchQuery}
         onFilterChange={setFilterSubscriptionId}
       />
-
       {selectedGroupIds.size > 0 && (
         <BulkActionsBar
           selectedCount={selectedGroupIds.size}
@@ -201,7 +140,6 @@ export function GroupsPage() {
           onBulkDelete={handleBulkDelete}
         />
       )}
-
       <GroupsTable
         groups={filteredGroups}
         subscriptions={subscriptions}
