+++ extension/entrypoints/dashboard/components/JobViewer.tsx
@@ -1,5 +1,4 @@
 import { useMemo, useState } from 'react';
-import { Button } from '@/components/ui/button';
 import {
   useCancelJob,
   useDeleteJob,
@@ -7,57 +6,9 @@ import {
   useResumeJob,
   useStartJob,
 } from '@/lib/hooks/storage/useJobs';
-import { useLogs } from '@/lib/hooks/storage/useLogs';
-import type { JobStatus, ScrapeJob } from '@/lib/types';
 import { LoadingSpinner } from './LoadingSpinner';
-
-function formatTimestamp(timestamp: number) {
-  return new Date(timestamp).toLocaleString();
-}
-
-const STATUS_BADGES: Record<JobStatus, string> = {
-  pending: 'bg-gray-500',
-  running: 'bg-blue-500',
-  paused: 'bg-yellow-500',
-  completed: 'bg-green-500',
-  failed: 'bg-red-500',
-  cancelled: 'bg-gray-500',
-};
-
-interface JobViewerHeaderProps {
-  hasActiveJob: boolean;
-  isPending: boolean;
-  isCancelling: boolean;
-  onStartJob: () => void;
-}
-
-function JobViewerHeader({
-  hasActiveJob,
-  isPending,
-  isCancelling,
-  onStartJob,
-}: JobViewerHeaderProps) {
-  return (
-    <div className="bg-white rounded-lg shadow p-6">
-      <div className="flex items-center justify-between">
-        <div>
-          <h2 className="text-xl font-semibold">Scraping Jobs</h2>
-          <p className="text-sm text-gray-600 mt-1">
-            Manage background scraping jobs for all enabled groups
-          </p>
-        </div>
-        <Button
-          onClick={onStartJob}
-          disabled={isPending || hasActiveJob || isCancelling}
-          variant="primary"
-          className="px-6 font-medium"
-        >
-          {isPending ? 'Starting...' : 'Start New Job'}
-        </Button>
-      </div>
-    </div>
-  );
-}
+import { JobCard, STATUS_BADGES } from './JobCard';
+import { JobViewerHeader } from './JobViewerHeader';
 
 export function JobViewer() {
   const jobsQuery = useJobs();
@@ -69,9 +20,8 @@ export function JobViewer() {
   const [expandedJobId, setExpandedJobId] = useState<string | null>(null);
   const [showLogsForJob, setShowLogsForJob] = useState<string | null>(null);
 
-  const jobs = jobsQuery.data ?? [];
+  const jobs = useMemo(() => jobsQuery.data ?? [], [jobsQuery.data]);
 
-  // Separate active and completed jobs
   const { activeJob, completedJobs } = useMemo(() => {
     const active = jobs.find(
       (j) => j.status === 'running' || j.status === 'paused'
@@ -85,7 +35,6 @@ export function JobViewer() {
       )
       .sort((a, b) => (b.completedAt || 0) - (a.completedAt || 0))
       .slice(0, 3);
-
     return { activeJob: active, completedJobs: completed };
   }, [jobs]);
 
@@ -114,12 +63,8 @@ export function JobViewer() {
     if (confirm('Are you sure you want to delete this job?')) {
       deleteJobMutation.mutate(jobId, {
         onSuccess: () => {
-          if (expandedJobId === jobId) {
-            setExpandedJobId(null);
-          }
-          if (showLogsForJob === jobId) {
-            setShowLogsForJob(null);
-          }
+          if (expandedJobId === jobId) setExpandedJobId(null);
+          if (showLogsForJob === jobId) setShowLogsForJob(null);
         },
       });
     }
@@ -153,8 +98,6 @@ export function JobViewer() {
         isCancelling={cancelJobMutation.isPending}
         onStartJob={handleStartJob}
       />
-
-      {/* Active Job Section */}
       {activeJob && (
         <div className="bg-white rounded-lg shadow">
           <div className="p-6 border-b border-gray-200">
@@ -179,8 +122,6 @@ export function JobViewer() {
           />
         </div>
       )}
-
-      {/* Historical Jobs Section */}
       {completedJobs.length > 0 && (
         <div className="bg-white rounded-lg shadow">
           <div className="p-6 border-b border-gray-200">
@@ -203,8 +144,6 @@ export function JobViewer() {
           </div>
         </div>
       )}
-
-      {/* Empty State */}
       {!activeJob && completedJobs.length === 0 && (
         <div className="bg-white rounded-lg shadow p-12 text-center">
           <p className="text-gray-600 mb-4">No jobs yet</p>
@@ -215,278 +154,4 @@ export function JobViewer() {
       )}
     </div>
   );
-}
-
-interface JobCardProps {
-  job: ScrapeJob;
-  isExpanded: boolean;
-  showLogs: boolean;
-  onToggleExpand: (jobId: string) => void;
-  onToggleLogs: (jobId: string) => void;
-  onCancel: (jobId: string) => void;
-  onResume: (jobId: string) => void;
-  onDelete: (jobId: string) => void;
-}
-
-function JobCard({
-  job,
-  isExpanded,
-  showLogs,
-  onToggleExpand,
-  onToggleLogs,
-  onCancel,
-  onResume,
-  onDelete,
-}: JobCardProps) {
-  const progressPercentage =
-    job.totalGroups > 0
-      ? Math.round((job.currentGroupIndex / job.totalGroups) * 100)
-      : 0;
-
-  return (
-    <div className="p-6">
-      {/* Job Summary */}
-      <div className="flex items-start justify-between">
-        <div className="flex-1">
-          <div className="flex items-center gap-3 mb-3">
-            <span
-              className={`px-2 py-1 rounded text-xs font-semibold text-white uppercase ${STATUS_BADGES[job.status]}`}
-            >
-              {job.status}
-            </span>
-            <span className="text-sm text-gray-600">
-              {formatTimestamp(job.createdAt)}
-            </span>
-            <span className="text-sm text-gray-600">
-              Duration:{' '}
-              {job.startedAt
-                ? formatDuration(
-                    job.startedAt,
-
-                    job.completedAt || undefined
-                  )
-                : 'Not started'}
-            </span>
-          </div>
-
-          {/* Progress Bar (for running/paused jobs) */}
-          {(job.status === 'running' || job.status === 'paused') && (
-            <div className="mb-4">
-              <div className="flex items-center justify-between text-sm mb-2">
-                <span className="font-medium">
-                  Progress: {job.currentGroupIndex} / {job.totalGroups} groups
-                </span>
-                <span className="text-gray-600">{progressPercentage}%</span>
-              </div>
-              <div className="w-full bg-gray-200 rounded-full h-2">
-                <div
-                  className="bg-blue-600 h-2 rounded-full transition-all"
-                  style={{ width: `${progressPercentage}%` }}
-                />
-              </div>
-            </div>
-          )}
-
-          {/* Stats */}
-          <div className="flex gap-6 text-sm">
-            <div>
-              <span className="text-gray-600">Total:</span>{' '}
-              <span className="font-medium">{job.totalGroups}</span>
-            </div>
-            <div>
-              <span className="text-gray-600">Success:</span>{' '}
-              <span className="font-medium text-green-600">
-                {job.successCount}
-              </span>
-            </div>
-            <div>
-              <span className="text-gray-600">Failed:</span>{' '}
-              <span className="font-medium text-red-600">
-                {job.failedCount}
-              </span>
-            </div>
-          </div>
-
-          {/* Error Message (if failed) */}
-          {job.error && (
-            <div className="mt-3 p-3 bg-red-50 border border-red-200 rounded text-sm text-red-700">
-              <strong>Error:</strong> {job.error}
-            </div>
-          )}
-        </div>
-
-        {/* Action Buttons */}
-        <div className="flex gap-2 ml-4">
-          {job.status === 'running' && (
-            <Button
-              type="button"
-              onClick={() => onCancel(job.id)}
-              variant="destructive"
-            >
-              Cancel
-            </Button>
-          )}
-          {(job.status === 'paused' || job.status === 'failed') && (
-            <Button
-              onClick={() => onResume(job.id)}
-              variant="primary"
-              size="sm"
-            >
-              Resume
-            </Button>
-          )}
-          {(job.status === 'completed' ||
-            job.status === 'failed' ||
-            job.status === 'cancelled') && (
-            <Button
-              onClick={() => onDelete(job.id)}
-              variant="destructive"
-              size="sm"
-            >
-              Delete
-            </Button>
-          )}
-          <Button
-            onClick={() => onToggleExpand(job.id)}
-            variant="secondary"
-            size="sm"
-          >
-            {isExpanded ? 'Hide Details' : 'Show Details'}
-          </Button>
-        </div>
-      </div>
-
-      {isExpanded && (
-        <div className="mt-6 border-t border-gray-200 pt-6 space-y-4">
-          <GroupResultsList groupResults={job.groupResults} />
-          <JobLogsSection
-            jobId={job.id}
-            showLogs={showLogs}
-            onToggleLogs={onToggleLogs}
-          />
-        </div>
-      )}
-    </div>
-  );
-}
-
-function GroupResultsList({
-  groupResults,
-}: {
-  groupResults: ScrapeJob['groupResults'];
-}) {
-  return (
-    <div>
-      <h4 className="font-medium mb-3">Group Results</h4>
-      <div className="space-y-2 max-h-96 overflow-y-auto">
-        {groupResults.map((result, index) => (
-          <div
-            key={result.groupId}
-            className="flex items-center gap-3 p-3 bg-gray-50 rounded text-sm"
-          >
-            <span className="text-gray-500 w-8">{index + 1}.</span>
-            {result.status === 'success' && (
-              <span className="text-green-600 text-lg">✓</span>
-            )}
-            {result.status === 'failed' && (
-              <span className="text-red-600 text-lg">✗</span>
-            )}
-            {result.status === 'pending' && (
-              <span className="text-gray-400 text-lg">⏳</span>
-            )}
-            {result.status === 'skipped' && (
-              <span className="text-gray-400 text-lg">⊘</span>
-            )}
-            <span className="flex-1 font-medium">{result.groupName}</span>
-            {result.postsScraped !== undefined && (
-              <span className="text-gray-600">{result.postsScraped} posts</span>
-            )}
-            {result.error && (
-              <span className="text-red-600 text-xs">{result.error}</span>
-            )}
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-interface JobLogsSectionProps {
-  jobId: string;
-  showLogs: boolean;
-  onToggleLogs: (jobId: string) => void;
-}
-
-function JobLogsSection({
-  jobId,
-  showLogs,
-  onToggleLogs,
-}: JobLogsSectionProps) {
-  const logsQuery = useLogs(showLogs ? { jobId } : undefined);
-  const logs = logsQuery.data ?? [];
-
-  return (
-    <div>
-      <Button onClick={() => onToggleLogs(jobId)} variant="secondary" size="sm">
-        {showLogs ? 'Hide Logs' : 'Show Logs'}
-      </Button>
-      {showLogs && (
-        <div className="border-t border-gray-200 pt-4 mt-4">
-          <h4 className="font-medium mb-3">Job Logs</h4>
-          {logsQuery.isLoading ? (
-            <div className="text-center py-4">
-              <LoadingSpinner />
-            </div>
-          ) : logs.length === 0 ? (
-            <p className="text-gray-500 text-sm">No logs for this job</p>
-          ) : (
-            <div className="max-h-96 overflow-y-auto space-y-2 font-mono text-xs">
-              {logs.map((log) => (
-                <div
-                  key={log.id}
-                  className="p-2 bg-gray-50 rounded border border-gray-200"
-                >
-                  <div className="flex items-start gap-2">
-                    <span className="text-gray-500 whitespace-nowrap">
-                      {new Date(log.timestamp).toLocaleTimeString()}
-                    </span>
-                    <span
-                      className={`px-1.5 py-0.5 rounded text-xs font-semibold uppercase ${
-                        log.level === 'error'
-                          ? 'bg-red-500 text-white'
-                          : log.level === 'warn'
-                            ? 'bg-yellow-500 text-white'
-                            : log.level === 'info'
-                              ? 'bg-blue-500 text-white'
-                              : 'bg-gray-500 text-white'
-                      }`}
-                    >
-                      {log.level}
-                    </span>
-                    <span className="flex-1">{log.message}</span>
-                  </div>
-                  {log.context && Object.keys(log.context).length > 0 && (
-                    <pre className="mt-1 text-xs text-gray-600 overflow-x-auto">
-                      {JSON.stringify(log.context, null, 2)}
-                    </pre>
-                  )}
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function formatDuration(
-  startedAt: number | null,
-  endTime: number = Date.now()
-) {
-  if (!startedAt) return 'Not started';
-  const durationMs = endTime - startedAt;
-  const minutes = Math.floor(durationMs / 60000);
-  const seconds = Math.floor((durationMs % 60000) / 1000);
-  return `${minutes}m ${seconds}s`;
 }
