+++ extension/entrypoints/dashboard/components/LogViewer.tsx
@@ -1,38 +1,9 @@
 import { useMemo, useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { useClearLogs, useLogs } from '@/lib/hooks/storage/useLogs';
-import type { LogEntry, LogLevel, LogSource } from '@/lib/types';
+import type { LogLevel, LogSource } from '@/lib/types';
 import { LoadingSpinner } from './LoadingSpinner';
-
-const LOG_LEVEL_COLORS: Record<LogLevel, string> = {
-  debug: 'bg-gray-100 text-gray-700 border-gray-300',
-  info: 'bg-blue-100 text-blue-700 border-blue-300',
-  warn: 'bg-yellow-100 text-yellow-700 border-yellow-300',
-  error: 'bg-red-100 text-red-700 border-red-300',
-};
-
-const LOG_LEVEL_BADGES: Record<LogLevel, string> = {
-  debug: 'bg-gray-500',
-  info: 'bg-blue-500',
-  warn: 'bg-yellow-500',
-  error: 'bg-red-500',
-};
-
-function formatTime(timestamp: number): string {
-  const date = new Date(timestamp);
-  return date.toLocaleTimeString('en-US', {
-    hour12: false,
-    hour: '2-digit',
-    minute: '2-digit',
-    second: '2-digit',
-    fractionalSecondDigits: 3,
-  });
-}
-
-function formatContext(context?: Record<string, unknown>): string {
-  if (!context || Object.keys(context).length === 0) return '';
-  return JSON.stringify(context, null, 2);
-}
+import { LogEntryRow } from './LogEntryRow';
 
 export function LogViewer() {
   const logsQuery = useLogs();
@@ -42,7 +13,7 @@ export function LogViewer() {
   const [sourceFilter, setSourceFilter] = useState<LogSource | 'all'>('all');
   const [autoScroll, setAutoScroll] = useState(true);
 
-  const logs = logsQuery.data ?? [];
+  const logs = useMemo(() => logsQuery.data ?? [], [logsQuery.data]);
 
   const filteredLogs = useMemo(() => {
     return logs.filter((log) => {
@@ -72,11 +43,9 @@ export function LogViewer() {
 
   return (
     <div className="flex flex-col h-full">
-      {/* Filters and Controls */}
       <div className="bg-white border-b border-gray-200 p-4 space-y-4">
         <div className="flex flex-wrap gap-4 items-center justify-between">
           <div className="flex gap-4 items-center">
-            {/* Level Filter */}
             <div className="flex items-center gap-2">
               <label htmlFor="level-filter" className="text-sm font-medium">
                 Level:
@@ -96,8 +65,6 @@ export function LogViewer() {
                 <option value="error">Error</option>
               </select>
             </div>
-
-            {/* Source Filter */}
             <div className="flex items-center gap-2">
               <label htmlFor="source-filter" className="text-sm font-medium">
                 Source:
@@ -117,8 +84,6 @@ export function LogViewer() {
                 <option value="dashboard">Dashboard</option>
               </select>
             </div>
-
-            {/* Auto-scroll Toggle */}
             <label className="flex items-center gap-2 text-sm">
               <input
                 type="checkbox"
@@ -129,7 +94,6 @@ export function LogViewer() {
               Auto-scroll
             </label>
           </div>
-
           <div className="flex gap-2">
             <Button
               type="button"
@@ -141,8 +105,6 @@ export function LogViewer() {
             </Button>
           </div>
         </div>
-
-        {/* Stats */}
         <div className="flex gap-4 text-sm text-gray-600">
           <span>
             Total: <strong>{logs.length}</strong>
@@ -152,8 +114,6 @@ export function LogViewer() {
           </span>
         </div>
       </div>
-
-      {/* Log Entries */}
       <div className="flex-1 overflow-auto bg-gray-50 p-4">
         {filteredLogs.length === 0 ? (
           <div className="text-center py-8 text-gray-500">
@@ -169,40 +129,5 @@ export function LogViewer() {
         )}
       </div>
     </div>
-  );
-}
-
-function LogEntryRow({ log }: { log: LogEntry }) {
-  const [expanded, setExpanded] = useState(false);
-
-  return (
-    <button
-      type="button"
-      className={`w-full text-left border rounded p-3 ${LOG_LEVEL_COLORS[log.level]} hover:opacity-90 transition-opacity`}
-      onClick={() => setExpanded(!expanded)}
-    >
-      <div className="flex items-start gap-3">
-        <span className="text-gray-500 whitespace-nowrap">
-          {formatTime(log.timestamp)}
-        </span>
-        <span
-          className={`px-2 py-0.5 rounded text-xs font-semibold text-white uppercase ${LOG_LEVEL_BADGES[log.level]}`}
-        >
-          {log.level}
-        </span>
-        <span className="px-2 py-0.5 rounded text-xs font-semibold bg-gray-700 text-white uppercase">
-          {log.source}
-        </span>
-        <span className="flex-1">{log.message}</span>
-        {log.context && Object.keys(log.context).length > 0 && (
-          <span className="text-gray-400 text-xs">{expanded ? '▼' : '▶'}</span>
-        )}
-      </div>
-      {expanded && log.context && Object.keys(log.context).length > 0 && (
-        <pre className="mt-2 p-2 bg-black bg-opacity-10 rounded text-xs overflow-x-auto">
-          {formatContext(log.context)}
-        </pre>
-      )}
-    </button>
   );
 }
