+++ extension/eslint.config.js
@@ -12,4 +12,20 @@ export default [
       'no-console': 'off',
     },
   },
+  {
+    // TanStack Virtual returns functions that React Compiler flags as incompatible;
+    // this is a known false positive for virtualizer hooks.
+    files: ['**/VirtualPostList.tsx'],
+    rules: {
+      'react-hooks/incompatible-library': 'off',
+    },
+  },
+  {
+    // Test files: relax function-length rules (test suites can be long)
+    files: ['**/*.test.ts', '**/*.test.tsx'],
+    rules: {
+      'max-lines-per-function': 'off',
+      'max-lines': 'off',
+    },
+  },
 ];
